Added Pages

+++ Telemeal/Pages/EmployeeLogin_Page.xaml.cs
@@ -47,7 +47,7 @@ namespace Telemeal.Pages
             bool admin = false;
             while (reader.Read())
             {
-                ADMINID = ((int)reader["ID"]).ToString();
+                ADMINID = (reader["ID"]).ToString();
                 ADMINNAME = ((string)reader["name"]);
                 if (EmployeeID.Password.Equals(ADMINID) && EmployeeName.Text.Equals(ADMINNAME))
                 {
